A finished battle
Just finished this.  I tested several scenarios and the battle seems to be running smoothly now for the most part.  One notable issue is the item usage.  The item is used and the hp is reflected properly, but the item is not being removed from the printable list (resulting in an infinite number of hp-restoring items).  This needs to be rectified, but I have done all that I can afford tonight.  Everything else, from my tests, is in working order.